sched: Reduce overestimating rq->avg_idle
When updating avg_idle, if the delta exceeds some max value, then
avg_idle
gets set to the max, regardless of what the previous avg was. This can
cause
avg_idle to often be overestimated.

This patch modifies the way we update avg_idle by always updating it
with the
function call to update_avg() first. Then, if avg_idle exceeds the max,
we set
it to the max.

Signed-off-by: Jason Low <jason.low2@hp.com>
Reviewed-by: Rik van Riel <riel@redhat.com>

Signed-off-by: Paul Reioux <reioux@gmail.com>